Cleanup: if conditions following failed .get()
.get() by default will return None if the key is not found in the
dictionary.

Thus if the following does not find the key then x == None.

x = dict.get('something', None)

This requires any conditional to test against None as follows:
if x is not None:

But in place in the code we're doing:
if x:

This correct that and simplifies some conditionals if needed.