Correctly display sort by priority in editor
In cases where a virtual folder location is a TP root, and we
are translating a subdirectory with files included withing a virtual
folder, but this subdirectory is not any vfolder location, we were
failing to tell if it had any VirtualFolderTreeItem.

This fix works fine since this class only allows to translate
directory paths within TPs.

Fixes #4504.