EditorTextarea: allow removing newline symbols
This symbol is special to the rest, because the characters displayed in the
textarea are not pure replacements, but symbol + character. Therefore, removing
a newline symbol would bring that back to the user's eye, as the font filter
would find an actual `\n which wasn't removed, and add the symbol for it.

For this reason, we listen for `keydown` events in editor textareas and remove
the symbol + character ourselves when we detect a newline symbol wants to be
removed.