ENChecker: fix running checks via `calculate_checks`
Due to the confusing double code-path for running checks, checkers are locked
into the setup being done in `run_given_filters()` if they want to remain
compatible with running individual checks from the CLI.

At the same time, note `calculate_checks` never passes an actual unit to the
check runners for performance reasons, but a trimmed-down version of it, and
checkers need to therefore implicitly rely on a CLI command's implementation
details.

These extra concerns should be addressed when dealing with #4553.