RawFontTextarea: consolidate all logic in a single component
This commit puts all raw font logic in a single component, which moves the state
away to parent components, simplifies the overall component tree and also allows
to get rid of some part of the ugly code being used in order to allow the
DOM-based editor to deal with the React-based editor state.

Setting values as props from the outside world has also been made more explicit,
by accepting an `overrideValues` prop instead of `values`, which clearly exposes
the intent and the expected side-effect.

The remaining not-so-beautiful parts have been explicitly commented in the code
to allow an easy removal when the whole editor state is managed via a
components.