Editor: use wrapper for textarea
Since in the near-future more elements will need to be combined with the
textarea (live previews etc.), the wrapper provides a convenient way to host all
of these while providing a way to nicely style everything.

Implementation-wise, reacting to focus/blur events is needed because there's no
way to target parent elements in pure CSS (note this is the same solution we use
for the search box).