Mousetrap: always prevent default browser behavior
Mousetrap doesn't offer a default behavior or global hook to disallow events
from propagating, therefore this is left for callbacks to do it explicitly.
That's the case for Pootle: we want to hijack shortcuts entirely, and not bubble
up the event, hence avoiding potential side-effects.

As an example, on Windows machines, pressing Ctrl+Up/Down in the editor would
trigger Pootle's logic AND the native behavior, resulting in potential scrolls
as the event could be handled after the editor changed its layout.