jsi18n: ensure default locale is built
The `get_current_language()` function, which is used in the main _layout.html_
template to build the URL in conjunction with the `{% statici18n %}` template
tag, will always return `settings.LANGUAGE_CODE` when there's no matching locale
for the user's preference or when `USE_I18N = False`. This means that in such
cases the generated URL will have `en-us` as the language code.

On the other hand, the `compilejsi18n` command reads the list of locales to
process from `settings.LANGUAGE_CODES`, which doesn't include `en-us` in
Pootle's case, and hence any references to `/static/locale/en-us/<filename.js>`
will be broken.

The existing band-aid was to provide the `-l en-us` flag to the `compilejsi18n`
command, where `get_javascript_catalog()` would try to locate a `en-us`
directory, fail to do so and fall back to the default `settings.LANGUAGE_CODE`
locale. This requires having the knowledge about all these nuances, special
casing and repeating code which we can avoid with a simple symbolic link.

Ideally this should be handled transparently by django-statici18n, so that when
it sees `USE_I18N = False`, it processes `settings.LANGUAGE_CODE` (and only that
locale, with no need to go through the entire list of available locales), and
everything would just work.