RawFontTextarea: do not report changes while in composition
Typing characters while having an IME open fires change events, and this
triggers the caret repositioning logic after the contents have been re-rendered
and flushed to the DOM. This generates a bunch of bugs for scenarios where an
IME is used, making the experience unusable.

React itself should take care of this but since it doesn't yet (there's a bug on
file for that, 3926), we need to _manually_ abort reporting any changes while
composition is open.