Editor: fix setting values out of the component tree
The textarea values are now set imperatively; we should do this because the
textarea value is entirely uncontrolled, so it's the only reliable way to change
the textarea's internal state. Besides that, this has an upside in the form of
removing a bunch of hairy component code.

Note the glue code to make this happen is ad-hoc and far from any kind of
beauty, although it gets the job properly done for now. As an improvement, we
should drill holes through the component tree by applying some unidirectional
data flow pattern.