Req: freshen requires.io filters
Note a rply from requires.io in this regard:

We actually support compatible versions (~=), it might just not work
exactly as you expect.

For example if you specify Django~=1.9.10, you will be green as long as
the latest Django release available on pypi is <1.10.
Since Django 1.10.x has been released, you need to specify a filter,
which can be any of the following:
rq.filter: >=1.9,<1.10
rq.filter: ~=1.9.10
rq.filter: <1.10

Note that you can use ~= in the filter too, the syntax is the same as
the syntax of the requirement.