Do not set cookie on response and use the old cookie name
There is no need to be sending back the cookie since it is
now only used by the client to alter the openness sidebar
status.

Also the cookie is not trusted, and the sidebar openness
is kept on the session.

Since now we actively remove the cookie from the response,
if by chance any client sends a big old cookie it will be
removed forever.

Fixes #5638 and #5637.