Contact: Do not expose report_email report error form
Note that unit is retrieved after self.request is populated. This
fixes sending report string error emails with broken subject (no unit
number, no language, no project).

It is unperformant to retrieve it several times, but it is safer
to do it like this as it was not possible to identify any method
that is called always after `as_view()` (which is where self.request
is being set) but before the methods that need the unit.