Update to work with apache 2.4 out of the box
This update allows static assets to be displayed out of the box when using Apache 2.4.  Apache 2.2 users will receive an error (among the other errors they may already have to fix), which will lead them directly to the comment.

Of course, this might not be the best approach, as the PR process may bring to light.