Docs: Break out language codes mapping to its own subsubsection
Also minor rewording and emphasized this is optional.

The reason to break out is to allow pointing people to this specific
topic and also reduce the length of the Creating a project instructions.

Note we keep the short instructions on language codes mapping in the
"Initialize new TP" instructions because we want to avoid that reading
the docs becames a pinball game.